Horizontally launched projectile motion
Data, analysis, and plotted results from analysis of object being launched horizontally off a table. Data is initially read from a text file for parsing/cleaning. The data is then place in a pandas data frame for analysis and plotting. A scikitlearn linear regression model is built for the data and the predicted line is added to the plot. The slope of the linear regression can be used to estimate the acceleration due to gravity since the vertical displacement of the ball bearing is known.